Remove context changes, focus more on http.HandlerFunc
This updates the primary API to no longer pass contexts explicitly but
instead encourage the use of http.Request's context. This makes the API
slightly more compatible with more generic stdlib focused middleware.

Additionally, this fixes the argument order for middleware and internal
code to be more consistent with the stdlib.