Experiment with using generic Data instead of generic "action"
This replaces the currently overridable "action" type with a generic
`Request[Data any]` type. This allows the user to define their own data type
that is passed to the handlers without having to re-instantiate or use
embedding to add additional fields to the action type.

This results in behavior no longer being defined on the action type, but
instead should accept either the request or the data type directly.

Eventually the response writer will be replaced by having handlers
return a response object that can be used to write the response and
improve type safety a tiny bit.